+++ CHANGELOG.md
@@ -4,9 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.1.15 - 02-26-2021
+* larva-js - Fix sideskin ads
+
+## 0.1.14 - 02-25-2021
+* larva-svg - Update pmc logo
+
+## 0.1.13 - 02-10-2021
+* larva-tokens - Add The Hollywood Reporter (THR) design tokens.
 
 ## 0.1.12 - 02-08-2021
 * Typo - Fix Variable typo
+
 ## 0.1.11 - 02-08-2021
 * Accessibility - Add aria-label attribute to c-link component
 * PMCP-2796: Add focus trap to mega menu modal